move ci logic into module init file
